Add basic backend interaction w/ mySQL DB Due to the change in structure from separate backend/frontend to changed package name in backend and imported it into frontend. Backend now features basic interaction with a mySQL database in the form of opening a connection to the database and pinging it.
dSN stands for dataSourceName and that is what you call the "address" of the database that you must "dial" to connect to the database.
The simplest dSN is in the form of [user]:[password]@/[databaseName].
[user] and [password] are what you use to log in to phpmyadmin. [databaseName] is "pricosha" for our purposes.

As we will be working on different local mySQL servers, we will need to enter different info for that to work.
So, I wrote a simple config file reader that takes a .json file with your login info and the dbName (in case we change it) and creates the dSN from that.
Currently, the file must be named config.json and must be located in the same directory as the backend (i.e. /backend).
The file must look like this: https://i.imgur.com/SkhnT1u.png

Additionally, I added a .gitignore file to ignore .json files in our repositories as I didn't want to track the config.json file.
I may change this implementation if I think of something better.