Implement base feature: add friend behavior
This commit is dedicate to Jayson Isaac and git respectively.
Adds frontend behavior for users viewing the friend groups they're in
and allows them to proceed to a FriendGroup specific page where they
can add friends to the FriendGroup via a form.
Adds backend behavior to handle this data manipulation and respond to
requests from frontend whenever it needs to validate a users data.

Specific mentions:
     backend/add_friend_related.go contains functions related
     specifically to the functionality of the add friend behavior
     backend/friend_group.go contains functions that have an ability to
     be reused related to FriendGroups in general