Add role recognition on access of FriendGroup
Reworked GetBelongFriendGroup() to return []*BFGDataElement which is
a struct containing an int for role and FriendGroup struct. This was
necessary as links for going to the manage privilege page are not
meant to appear for members but for mods and owners.

Built out the /managePrivilege page to include a list of users in
various roles and the first steps towards promotion and demotion.

This can be tested by using first importing the the new pricosha.sql
file either via phpmyadmin or by calling the source command from
within a mysql command line interface. Afterwards, the user AA@nyu.edu
is an admin of the group family and as such can manage the privileges
of that group. She is also an admin of "nomembersnomods" which should
display the correct templating for lack of a certain role in a group.

Additionally, logging in as BB@nyu.edu allows you to manage the
privileges of AA@nyu.edu's family group as he is now a mod. Notice
how the links to demote and promote members do not appear for
BB@nyu.edu as he is only a mod.

Finally, logging in as CC@nyu.edu, who is a member of AA@nyu.edu's
family group, shows you no manage privileges buttons as she is only a
member.