Fix the index out of range error on upload
On the content item upload form, we show the user checkboxes so they
can select which friend groups they want to share their content with. On
the backend, we use the `value` attribute of those checkboxes to
determine what email and friend group name to add to the Share table.

For some reason, `{{ .OwnerEmail }}` was being rendered as a comment in
upload.html. This was causing a "index out of range" error when we were
trying to parse the checkbox values. By removing the space before the
template variable, it is no longer rendered as a comment.

Resolves: Issue #39