Add demotion and promotion of users.
Added changePrivilegeHandler() to handle demotion and promotion
requests as well as its sister function on the backend
ChangePrivilege() which executes them.

Reworked managePrivilege to work off of post data instead of query
parameters.

You can test this feature by signing in as AA@nyu.edu, going to the
manage privileges page for her "family" group, and smashing those
demote and promote buttons to change the roles.