User Privileges (#41)
* Add role field to Belong table in SQL.

1. Added a simple integer field to the Belong table in the pricosha.sql
     that defaults to 0. The behavior of that field is: 0 = member,
     1 = mod, 2 = admin.
2. Fixed situations in code where Belong table is queried and they don't
     account for the new field
3. Prettied the pricosha.sql file in general to user batch statements
     and print the label of the query before running it to identify
     where problems occured easier.

* Add handling of /managePrivilege requests

Add a managePrivilegesHandler on the frontend that handles all requests
to the /managePrivileges domain and the initial errors and serves a
skeletal HTML file of the planned page.

Add file to backend to put managing privilege related functions in.

Misc style changes.

* Add role recognition on access of FriendGroup

Reworked GetBelongFriendGroup() to return []*BFGDataElement which is
a struct containing an int for role and FriendGroup struct. This was
necessary as links for going to the manage privilege page are not
meant to appear for members but for mods and owners.

Built out the /managePrivilege page to include a list of users in
various roles and the first steps towards promotion and demotion.

This can be tested by using first importing the the new pricosha.sql
file either via phpmyadmin or by calling the source command from
within a mysql command line interface. Afterwards, the user AA@nyu.edu
is an admin of the group family and as such can manage the privileges
of that group. She is also an admin of "nomembersnomods" which should
display the correct templating for lack of a certain role in a group.

Additionally, logging in as BB@nyu.edu allows you to manage the
privileges of AA@nyu.edu's family group as he is now a mod. Notice
how the links to demote and promote members do not appear for
BB@nyu.edu as he is only a mod.

Finally, logging in as CC@nyu.edu, who is a member of AA@nyu.edu's
family group, shows you no manage privileges buttons as she is only a
member.

* Add demotion and promotion of users.

Added changePrivilegeHandler() to handle demotion and promotion
requests as well as its sister function on the backend
ChangePrivilege() which executes them.

Reworked managePrivilege to work off of post data instead of query
parameters.

You can test this feature by signing in as AA@nyu.edu, going to the
manage privileges page for her "family" group, and smashing those
demote and promote buttons to change the roles.

* Make mods able to add friends.

Fix spelling error in GetUserFriendGroup()
Remove extraneous links to /friendgroups and /logout in main.html
Add login and role checking to /addFriend and /formAddFriend

* Make users with correct privileges able to unshare

Original posters and user with correct privileges over their group can
now unshare items from said group on the Content_Item page.

This can be tested by posting something as BB@nyu.edu to his family and
AA@nyu.edu's family, logging out, viewing that item as AA@nyu.edu,
logging back in as BB@nyu.edu and unsharing the item from AA@nyu.edu's
family and then finally logging back in as AA@nyu.edu and noticing that
it is gone.

Additionally, you can test original content poster privileges by
posting an item as CC@nyu.edu to AA@nyu.edu's family and then going to
the ContentItem page and unsharing it.

Additionally, you can test mod and admin privileges by undergoing the
previous steps but instead of unsharing as CC@nyu.edu, log in as either
BB@nyu.edu (mod) or AA@nyu.edu (admin) and unshare the item there.

Finally, you can test lack of privileges by undergoing the previous
posting steps but instead of unsharing as CC@nyu.edu, log in as
HH@nyu.edu and view the Content_Item to see the option to unshare is
not there.

* Make admins able to rename FriendGroup

This can be tested by signing in as BB@nyu.edu and going first to
the manage privilege page of AA@nyu.edu's family so see that he cannot
see the admin options (rename group and swap owner). After, visit the
manage privilege page of BB@nyu.edu's family, click on the new Rename
button, fill in a name, and click Rename! to rename the group.

Also added new sql file to test changing of primary keys.

* Make admin able to change owner of FriendGroup

Test with BB@nyu.edu
Option shouldn't appear to non-admins

Test edge cases:
	User doesn't exist (try: asdasda)
	User specified for swap already has group with the same name (try:
		AA@nyu.edu)

* Make admins able to delete FriendGroup

Test with BB@nyu.edu
Option shouldn't appear to non-admins

Test edge cases:
	FriendGroup doesn't exist (try: asdada)

* Integrate user privileges into delete friend

Adds logged in and role status checking to delete friend pages.

* Fix GetUserContent() query to use FG primary key

GetUserContent()'s query was fixed in the following way:
	Unsafe wildcard select was substituted
	WHERE clause that gets if a user is in friend group with
	primary key instead of just the friend group name

* Fix case difference RenameFG() bug

SQL is case insensitive as such renaming a friend group from "family"
to "Family" does not work in SQL and must be accounted for in the Go
code to prevent simple deletion of a FriendGroup on rename