Add frontend handling of voting
Can be tested by signing in as BB@nyu.edu and clicking on the "Do you
like apples?" ContentItem. BB@nyu.edu has already voted "Yes" so by
voting again, you will change the vote he has cast.

If you sign in as HH@nyu.edu and vote on the same item, a new vote will
be added to the votes.

Notable effect on the ecosystem:
	1. When users are removed from FriendGroups, votes should be removed
	from Vote. This will situation will need to be handled gracefully.
	Leaving Vote's "floating" is inelegant.
	It is of my opinion that this does not apply to when public content
	leaves the 24 hour window and the Vote should remain forever.

	2. Luckily, due to on delete cascade behavior, this does not need to
	be handled on the deletion of a poll.

Possible mods of improvement:
	1. Vote buttons next to options to skip the text field entirely
	2. Multi-vote (Messenger style)
	3. Unvote (Delete vote completely)
	4. Predefining options (impossible without redefining system)